services/auth-service/requirements.txt
flask==2.3.3
gunicorn==21.2.0
pyjwt==2.8.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
redis==5.0.1
pika==1.3.2
prometheus-flask-exporter==0.22.4
flask-limiter==3.5.0
marshmallow==3.20.1 